+++ src/Markdown/Extension/Node/TableCaption.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Jidaikobo\Markdown\Extension\Node;
+
+use League\CommonMark\Node\Block\AbstractBlock;
+
+final class TableCaption extends AbstractBlock
+{
+}